# agentauth/docs/openapi.yaml
openapi: 3.1.0
info:
  title: AgentAuth API
  version: 0.7.0
  description: |
    **AgentAuth** is a production-ready authentication and identity service for AI agents.

    ## Features
    - JWT-based authentication
    - Rate limiting (100 req/15min general, 10 req/15min auth)
    - Webhooks for events
    - Activity logging
    - Permission-based access control
    - **Persona System** — define and manage an agent's digital soul (v0.7.0)
    - **ZKP Anonymous Verification** — prove identity without revealing credentials (v0.7.0)
    - **Anti-Drift Vault** — detect behavioral drift and auto-revoke rogue agents (v0.7.0)

    ## Authentication
    Most endpoints require an API key in the `X-Api-Key` header:
    ```
    X-Api-Key: ag_sk_your_api_key_here
    ```

    ## Rate Limits
    - **General endpoints**: 100 requests per 15 minutes
    - **Authentication endpoints**: 10 requests per 15 minutes

    Rate limit info is included in response headers:
    - `X-RateLimit-Limit`: Max requests allowed
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Time until limit resets (Unix timestamp)

    ## Base URL
    - **Production**: `https://api.agentauth.dev`
    - **Local**: `http://localhost:3000`

  contact:
    name: AgentAuth Support
    email: umytbaynazarow754@gmail.com
    url: https://agentauth.dev
  license:
    name: MIT
    url: https://github.com/umyt-dev/Agent-Identity/blob/main/LICENSE
servers:
  - url: https://api.agentauth.dev
...
paths:
...
  /v1/agents/verify:
    post:
      summary: Verify agent credentials
      description: |
        Verify agent credentials (agent_id + API key) and receive agent details.

        **Rate limit:** 10 requests per 15 minutes
      operationId: verifyAgent
      tags:
        - Agents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentVerificationRequest'
            example:
              agent_id: ag_1a2b3c4d5e6f
              api_key: ag_sk_1a2b3c4d5e6f7g8h9i0j
      responses:
        '200':
          description: Agent verified successfully
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentVerificationResponse'
              example:
                valid: true
                agent_id: ag_1a2b3c4d5e6f
                name: CustomerSupportAgent
                tier: free
                permissions:
                  - read:messages
                  - write:responses
                status: active
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimitExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /v1/drift/{agent_id}/drift-score:
    get:
      summary: Get drift score
      description: |
        Retrieve the current drift score and trend for an agent. The score is
        a weighted average of metric deltas from baseline (0.0 = no drift, 1.0 = max drift).
      operationId: getDriftScore
      tags:
        - Drift
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Current drift score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftScoreResponse'
              example:
                agent_id: ag_1a2b3c4d5e6f
                drift_score: 0.12
                trend: improving
                last_ping_at: '2026-02-01T12:00:00.000Z'
                status: healthy
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/drift/{agent_id}/drift-history:
    get:
      summary: Get drift history
      description: |
        Retrieve paginated drift score history for an agent.
        Supports CSV export and single-metric filtering.
      operationId: getDriftHistory
      tags:
        - Drift
      parameters:
        - $ref: '#/components/parameters/AgentId'
        - name: limit
          in: query
          description: Maximum entries to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          description: Number of entries to skip
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: metric
          in: query
          description: Filter by a single metric name
          schema:
            type: string
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Drift score history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftHistoryResponse'
            text/csv:
              schema:
                type: string
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/drift/{agent_id}/drift-config:
    put:
      summary: Update drift configuration
      description: |
        Configure drift thresholds, metric weights, auto-revoke behavior,
        spike sensitivity, and baseline metrics for an agent.
      operationId: updateDriftConfig
      tags:
        - Drift
      parameters:
        - $ref: '#/components/parameters/AgentId'
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriftConfigRequest'
            example:
              drift_threshold: 0.7
              warning_threshold: 0.5
              auto_revoke: true
              spike_sensitivity: 2.0
              metric_weights:
                toxicity_score: 3.0
                hallucination_rate: 2.0
                error_rate: 1.0
              baseline_metrics:
                response_time: 0.3
                error_rate: 0.01
                toxicity_score: 0.02
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftConfigResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: Get drift configuration
      description: Retrieve the current drift configuration for an agent.
      operationId: getDriftConfig
      tags:
        - Drift
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Drift configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriftConfigResponse'
              example:
                agent_id: ag_1a2b3c4d5e6f
                drift_threshold: 0.7
                warning_threshold: 0.5
                auto_revoke: true
                spike_sensitivity: 2.0
                metric_weights:
                  toxicity_score: 3.0
                  hallucination_rate: 2.0
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ─── Webhooks ────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Agent API key
  parameters:
    AgentId:
      name: agent_id
      in: path
      description: Unique agent identifier
      required: true
      schema:
        type: string
        pattern: '^ag_[a-zA-Z0-9]+$'
      example: ag_1a2b3c4d5e6f
...
  headers:
    X-RateLimit-Limit:
      description: Maximum number of requests allowed in the current window
      schema:
        type: integer
      example: 10
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window
      schema:
        type: integer
      example: 7
    X-RateLimit-Reset:
      description: Unix timestamp when the rate limit resets
      schema:
        type: integer
      example: 1738329600
  schemas:
...
    AgentVerificationRequest:
      type: object
      required: [agent_id, api_key]
      properties:
        agent_id:
          type: string
        api_key:
          type: string
    AgentVerificationResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        agent_id:
          type: string
        name:
          type: string
        tier:
          type: string
          enum: [free, pro, enterprise]
        permissions:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        error:
          type: string
...
    DriftScoreResponse:
      type: object
      properties:
        agent_id:
          type: string
        drift_score:
          type: number
          nullable: true
          minimum: 0
          maximum: 1
        trend:
          type: string
          enum: [improving, stable, worsening]
          nullable: true
        last_ping_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [healthy, warning, critical]
    DriftHistoryResponse:
      type: object
      properties:
        history:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              drift_score:
                type: number
              metrics:
                type: object
              spikes:
                type: array
                items:
                  type: string
              created_at:
                type: string
                format: date-time
        total:
          type: integer
    DriftConfigRequest:
      type: object
      properties:
        drift_threshold:
          type: number
          minimum: 0
          maximum: 1
          description: Score above which agent is considered drifted
        warning_threshold:
          type: number
          minimum: 0
          maximum: 1
          description: Score above which a warning webhook fires (must be < drift_threshold)
        auto_revoke:
          type: boolean
          description: Auto-revoke agent when drift exceeds threshold
        spike_sensitivity:
          type: number
          minimum: 0
          exclusiveMinimum: true
          description: Standard deviations for spike detection (default 2.0)
        metric_weights:
          type: object
          additionalProperties:
            type: number
            minimum: 0
          description: Per-metric weights for drift score calculation
        baseline_metrics:
          type: object
          additionalProperties:
            type: number
          description: Expected baseline metric values
    DriftConfigResponse:
      type: object
      properties:
        agent_id:
          type: string
        drift_threshold:
          type: number
        warning_threshold:
          type: number
        auto_revoke:
          type: boolean
        spike_sensitivity:
          type: number
        metric_weights:
          type: object
        baseline_metrics:
          type: object
        updated_at:
          type: string
          format: date-time

    # ─── Webhook Schemas ─────────────────────────────────────────────────────
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
  responses:
    BadRequest:
      description: Bad request - validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed
            message: Invalid input data
            details:
              - field: email
                message: Must be a valid email address
    Unauthorized:
      description: Unauthorized - invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Invalid agent credentials
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: Resource not found
    RateLimitExceeded:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Rate Limit Exceeded
            message: Too many requests. Please try again later.
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal Server Error
            message: An unexpected error occurred